+++ superwall/src/main/java/com/superwall/sdk/dependencies/DependencyContainer.kt
@@ -193,6 +193,7 @@ class DependencyContainer(
 
         transactionManager = TransactionManager(
             storeKitManager = storeKitManager,
+            purchaseController = purchaseController,
             sessionEventsManager,
             activityProvider,
             factory = this