+++ superwall/src/main/java/com/superwall/sdk/store/transactions/RestoreType.kt
@@ -1,0 +1,8 @@
+package com.superwall.sdk.store.transactions
+
+import com.superwall.sdk.store.abstractions.transactions.StoreTransaction
+
+sealed class RestoreType {
+    data class ViaPurchase(val transaction: StoreTransaction?) : RestoreType()
+    object ViaRestore : RestoreType()
+}